Split compiler.test.ts into 17 focused test files
Mechanical split of compiler.test.ts (5,022 lines, ~215 tests) into
single-concern test files. No behavior changes — each describe block
moved to the appropriate file:

- analyzer.test.ts: signal/props/memo extraction
- jsx-to-ir.test.ts: element/event/expression transforms
- ir-*.test.ts: IR structure tests (provider, ternary, nullish, etc.)
- boolean-attributes.test.ts: isBooleanAttr, DOM property binding
- client-js-generation.test.ts: codegen patterns (largest, 66 tests)
- adapter-output.test.ts: template output patterns
- And 6 more behavioral test files

All 410 tests pass with 1,280 expect() calls (unchanged).
